Move 'count' to private folder
